Skip pom files ending up on the classpath.
Sometimes libraries are badly packaged and you end up with pom files on the
classpath. Before this patch in such cases depstar would bail out. Instead I'm
making an explicit check for pom files and just skip them as the case is
"harmless". If some similat situation comes up for other garbage on class paths
it can be added to this case.

My specific case is when netlib/all 1.1.2 ends up with a pom file on the class
path fwiw (pulled in via the fastmath clojure lib).